reverse moving outline from `Compat`
