--- mtags/src/main/scala-2.11/scala/meta/internal/pc/Outline.scala
@@ -1,9 +1,0 @@
-package scala.meta.internal.pc
-
-import scala.meta.pc.OutlineFiles
-
-trait Outline { this: MetalsGlobal =>
-  def runOutline(files: OutlineFiles): Unit = {
-    // no outline compilation for 2.11
-  }
-}
