+++ mtags/src/main/scala-2/scala/meta/internal/pc/MetalsGlobal.scala
@@ -62,8 +62,7 @@ class MetalsGlobal(
     with GlobalProxy
     with AutoImports
     with Keywords
-    with WorkspaceSymbolSearch
-    with Outline { compiler =>
+    with WorkspaceSymbolSearch { compiler =>
   hijackPresentationCompilerThread()
 
   val logger: Logger = Logger.getLogger(classOf[MetalsGlobal].getName)
